Consolidate bridgetree tests.
